requirements.txt
Django==4.1.7
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow==9.4.0
drf-spectacular==0.26.1